Split javascript into modules
