+++ components/hosted_media/src/static/index.html
@@ -9,6 +9,6 @@
     <div id="state">Väntar på koppling till server</div>
 
     <!-- Content will be automatically added by scripts here -->
-    <script src="./script.js"></script>
+    <script src="./script.js" type="module"></script>
   </body>
 </html>
